Adding seperate config files for each component being tested.
