+++ .gitignore
@@ -25,4 +25,4 @@ ghpages
 # Generated CSS files
 docs/css/fiori-fundamentals.css
 docs/css/fui-site.css
-test/visual-regression-tests/backstopConfigCI.json
+test/visual-regression-tests/config/backstopConfigCI.json
