+++ .travis.yml
@@ -13,7 +13,7 @@ script:
 jobs:
   include:
   - stage: Test
-    script: gulp test:visualCI
+    script: gulp test:generateCIConfig && test:visual
   - stage: Pre-release
     if: branch = master
     before_deploy:
