+++ ops/gulp/test.js
@@ -6,8 +6,8 @@ const backstop = require('backstopjs');
 const fs = require('fs');
 
 const localAppPort = 3030;
-const backstopConfigLocation = 'test/visual-regression-tests/backstopConfig.js';
-const backstopCIConfigLocation = 'test/visual-regression-tests/backstopConfigCI.json';
+const backstopConfigLocation = 'test/visual-regression-tests/config/backstopConfig.js';
+const backstopCIConfigLocation = 'test/visual-regression-tests/config/backstopConfigCI.json';
 
 // Starts the test server.  If the server port is already in use, it is assumed that the server is already running.
 gulp.task('server:start', function (cb) {	
@@ -89,9 +89,6 @@ gulp.task('test:reference', gulp.series('server:start', 'backstop:reference'));
 
 // Tests the current visual elements against the reference screenshots.
 gulp.task('test:visual', gulp.series('server:start', 'backstop:test'));
-
-// Tests the current visual elements against the reference screenshots in CI.
-gulp.task('test:visualCI', gulp.series('test:generateCIConfig', 'server:start', 'backstop:test'));
 
 // Approves any changes to the screenshots and promotes them to the reference data set.
 gulp.task('test:approve', () => backstop('approve', {config: backstopConfigLocation }));
