+++ package.json
@@ -48,6 +48,7 @@
     "gulp-sourcemaps": "^2.6.4",
     "htmlparser2": "^3.10.0",
     "ip": "^1.1.5",
+    "merge-json": "0.1.0-b.3",
     "node-sass": "^4.9.4",
     "nunjucks": "^3.1.3",
     "read-file": "^0.2.0",
