+++ test/visual-regression-tests/config/components/alerts.js
@@ -1,0 +1,21 @@
+module.exports = {
+    "label": "Alerts",
+    "urlSuffix": "/alert",
+    "selectors": [
+        "div.fd-alert",
+        "div.fd-alert.fd-alert--information",
+        "div.fd-alert.fd-alert--success",
+        "div.fd-alert.fd-alert--warning",
+        "div.fd-alert.fd-alert--error",
+        "div.fd-alert.fd-alert--dismissible",
+        "div.fd-alert.fd-alert--information.fd-alert--dismissible",
+        "div.fd-alert.fd-alert--success.fd-alert--dismissible",
+        "div.fd-alert.fd-alert--warning.fd-alert--dismissible",
+        "div.fd-alert.fd-alert--error.fd-alert--dismissible",
+        "div.fd-alert.fd-alert--dismissible[dir=rtl]",
+        "div.fd-alert.fd-alert--success[dir=rtl]",
+        "div.fd-alert.fd-alert--warning.fd-alert--dismissible[dir=rtl]",
+        "div.fd-alert.fd-alert--error.fd-alert--dismissible[dir=rtl]"
+      ],
+      "selectorExpansion": true  
+};
